Move mapRelativePositions to its own file
