+++ src/components/Text/Annotated/utils/createTextOffsets.ts
@@ -2,16 +2,6 @@ import { AnnoRepoAnnotation } from "../../../../model/AnnoRepoAnnotation.ts";
 import { BlockType, TextOffsets } from "../core";
 import { BroccoliRelativeAnno } from "../../../../model/Broccoli.ts";
 
-export function mapRelativePositions(
-  relativePositions: BroccoliRelativeAnno[],
-): Map<string, BroccoliRelativeAnno> {
-  const map = new Map<string, BroccoliRelativeAnno>();
-  for (const p of relativePositions) {
-    map.set(p.bodyId, p);
-  }
-  return map;
-}
-
 export function createTextOffsets(
   annotation: AnnoRepoAnnotation,
   relativePosition: BroccoliRelativeAnno,
